MOD: add initial test code
