--- EmployeeDB/.vs/EmployeeDB/project-colors.json
@@ -1,16 +1,0 @@
-{
-  "Version": 1,
-  "ProjectMap": {
-    "19fa302f-bf01-4d2b-8c90-019c54671051": {
-      "ProjectGuid": "19fa302f-bf01-4d2b-8c90-019c54671051",
-      "DisplayName": "Test_Project",
-      "ColorIndex": 0
-    },
-    "f3b037b3-550e-4694-a0a7-27e1ef5bf2b8": {
-      "ProjectGuid": "f3b037b3-550e-4694-a0a7-27e1ef5bf2b8",
-      "DisplayName": "EmployeeDB",
-      "ColorIndex": 1
-    }
-  },
-  "NextColorIndex": 2
-}
