+++ EmployeeDB/EmployeeDB/EmployeeDB.vcxproj.filters
@@ -24,6 +24,9 @@
     <ClCompile Include="Adder.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
+    <ClCompile Include="Mod.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Parser.h">
