+++ EmployeeDB/EmployeeDB/x64/Debug/EmployeeDB.log
@@ -1,2 +1,1 @@
-﻿  Adder.cpp
-  EmployeeDB.vcxproj -> C:\Users\yoonj\OneDrive\문서\C_Clean_Code\addMerge2\EmployeeDB\x64\Debug\EmployeeDB.exe
+﻿
