+++ EmployeeDB/Test_Project/x64/Debug/Test_Project.Build.CppClean.log
@@ -1,0 +1,26 @@
+c:\users\yoonj\onedrive\문서\c_clean_code\addmerge2\employeedb\test_project\x64\debug\test_project.pch
+c:\users\yoonj\onedrive\문서\c_clean_code\addmerge2\employeedb\test_project\x64\debug\vc143.pdb
+c:\users\yoonj\onedrive\문서\c_clean_code\addmerge2\employeedb\test_project\x64\debug\pch.obj
+c:\users\yoonj\onedrive\문서\c_clean_code\addmerge2\employeedb\test_project\x64\debug\test.obj
+c:\users\noh02\desktop\과제\c_clean_code\employeedb\test_project\x64\debug\test_project.pch
+c:\users\noh02\desktop\과제\c_clean_code\employeedb\test_project\x64\debug\vc143.pdb
+c:\users\noh02\desktop\과제\c_clean_code\employeedb\test_project\x64\debug\pch.obj
+c:\users\noh02\desktop\과제\c_clean_code\employeedb\test_project\x64\debug\test.obj
+c:\users\noh02\desktop\과제\c_clean_code\employeedb\x64\debug\gtest.pdb
+c:\users\noh02\desktop\과제\c_clean_code\employeedb\x64\debug\gtest_main.pdb
+c:\users\yoonj\onedrive\문서\c_clean_code\addmerge2\employeedb\x64\debug\test_project.exe
+c:\users\yoonj\onedrive\문서\c_clean_code\addmerge2\employeedb\x64\debug\test_project.ilk
+c:\users\yoonj\onedrive\문서\c_clean_code\addmerge2\employeedb\x64\debug\test_project.pdb
+c:\users\noh02\desktop\과제\c_clean_code\employeedb\x64\debug\test_project.ilk
+c:\users\noh02\desktop\과제\c_clean_code\employeedb\x64\debug\test_project.exe
+c:\users\noh02\desktop\과제\c_clean_code\employeedb\x64\debug\test_project.pdb
+c:\windows\pcdrm\dump\20220215_nfd01.log
+c:\windows\pcdrm\dump\20220215_nsccor03.log
+c:\users\noh02\desktop\과제\c_clean_code\employeedb\test_project\x64\debug\test_project.tlog\cl.command.1.tlog
+c:\users\noh02\desktop\과제\c_clean_code\employeedb\test_project\x64\debug\test_project.tlog\cl.read.1.tlog
+c:\users\noh02\desktop\과제\c_clean_code\employeedb\test_project\x64\debug\test_project.tlog\cl.write.1.tlog
+c:\users\noh02\desktop\과제\c_clean_code\employeedb\test_project\x64\debug\test_project.tlog\copylocal.read.1u.tlog
+c:\users\noh02\desktop\과제\c_clean_code\employeedb\test_project\x64\debug\test_project.tlog\copylocal.write.1u.tlog
+c:\users\noh02\desktop\과제\c_clean_code\employeedb\test_project\x64\debug\test_project.tlog\link.command.1.tlog
+c:\users\noh02\desktop\과제\c_clean_code\employeedb\test_project\x64\debug\test_project.tlog\link.read.1.tlog
+c:\users\noh02\desktop\과제\c_clean_code\employeedb\test_project\x64\debug\test_project.tlog\link.write.1.tlog
