+++ EmployeeDB/Test_Project/x64/Debug/Test_Project.exe.recipe
@@ -5,7 +5,7 @@
       <FullPath>C:\Users\yoonj\OneDrive\문서\C_Clean_Code\addMerge2\EmployeeDB\x64\Debug\EmployeeDB.exe</FullPath>
     </ProjectOutput>
     <ProjectOutput>
-      <FullPath>C:\Users\yoonj\OneDrive\문서\C_Clean_Code\addMerge2\EmployeeDB\x64\Debug\Test_Project.exe</FullPath>
+      <FullPath>C:\Users\noh02\Desktop\과제\C_Clean_Code\EmployeeDB\x64\Debug\Test_Project.exe</FullPath>
     </ProjectOutput>
   </ProjectOutputs>
   <ContentFiles />
