--- EmployeeDB/Test_Project/x64/Debug/Test_Project.tlog/Test_Project.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.30.30705:TargetPlatformVersion=10.0.19041.0:
-Debug|x64|C:\Users\yoonj\OneDrive\문서\C_Clean_Code\addMerge2\EmployeeDB\|
